put the deletion in the if

+++ controller/ForumController.php
@@ -231,16 +231,15 @@
 
             $postManager = new PostManager();
             $topicManager = new TopicManager();
-            var_dump($postManager);die;
-            $postManager->delete($id);
-
-            /*if(($postManager->findOneById($id)->getUser() == Session::getUser()) OR Session::isAdmin()){
+            
+            if(($postManager->findOneById($id)->getUser() == Session::getUser()) OR Session::isAdmin()){
+                $postManager->delete($id);
                 Session::addFlash("success","vous avez supprimer le message avec succès");
                 $this->redirectTo("forum","aTopic",$_GET['idTopic']);
             }
             else{
                 Session::addFlash("error","vous avez tentez de supprimez un message qui ne vous appartient pas !");
                 $this->redirectTo("home");
-            }*/
+            }
         }
     }
